+++ tests/tt_metal/tt_metal/dispatch/dispatch_event/test_events.cpp
@@ -6,7 +6,6 @@
 #include <fmt/base.h>
 #include <cstddef>
 #include <cstdint>
-#include "impl/dispatch/command_queue.hpp"
 #include <tt-metalium/event.hpp>
 #include <tt-metalium/host_api.hpp>
 #include <future>
