+++ tests/tt_metal/tt_metal/lightmetal/test_lightmetal.cpp
@@ -6,7 +6,6 @@
 
 // #include <fmt/base.h>
 // #include <stddef.h>
-// #include "impl/dispatch/command_queue.hpp"
 // #include <tt-metalium/host_api.hpp>
 // #include <tt-logger/tt-logger.hpp>
 // #include <tt-metalium/program.hpp>
