+++ tt_metal/impl/device/device.cpp
@@ -28,7 +28,6 @@
 
 #include "allocator.hpp"
 #include <tt_stl/assert.hpp>
-#include "dispatch/command_queue.hpp"
 #include "dispatch/command_queue_common.hpp"
 #include "common/core_assignment.hpp"
 #include "program/program_impl.hpp"
