+++ tt_metal/impl/lightmetal/lightmetal_replay_impl.cpp
@@ -14,7 +14,6 @@
 #include "env_lib.hpp"
 #include <tt-metalium/tt_metal.hpp>
 #include "trace/trace_buffer.hpp"
-#include "impl/dispatch/command_queue.hpp"
 #include <tt-metalium/device.hpp>
 #include "flatbuffer/program_types_from_flatbuffer.hpp"
 #include "flatbuffer/buffer_types_from_flatbuffer.hpp"
